--- batch-nft-reveal/app/components/load/index.ts
@@ -1,1 +1,0 @@
-export { Load } from './Load'
